added global notifications

+++ cogs/notifications.py
@@ -39,14 +39,11 @@ class Notifications(commands.Cog):
         if message.author.id == self.bot.user.id:
             return
 
-        keywords = db.get_keywords(message.guild.id)
+        keywords = db.get_keywords(message)
         if keywords is None:
             return
 
         for word, user_id in keywords:
-            if user_id == message.author.id:
-                continue
-
             pattern = re.compile(r'(?:^|\W){0}(?:$|\W)'.format(word), flags=re.IGNORECASE)
             if pattern.findall(message.content):
                 member = message.guild.get_member(user_id)
@@ -58,7 +55,7 @@ class Notifications(commands.Cog):
                 # create and send notification message
                 await send_notification(member, message, pattern)
 
-    @commands.group(case_insensitive=True)
+    @commands.group(case_insensitive=True, aliases=['noti'])
     async def notification(self, ctx):
         """Add keyword notifications on this server"""
         await util.command_group_help(ctx)
@@ -66,32 +63,45 @@ class Notifications(commands.Cog):
     @notification.command()
     async def add(self, ctx, *, keyword):
         """Add a notification"""
-        await ctx.message.delete()
+        dm = ctx.guild is None
+        if dm:
+            guild_id = 0
+        else:
+            await ctx.message.delete()
+            guild_id = ctx.guild.id
 
         check = db.query("SELECT * FROM notifications WHERE guild_id = ? and user_id = ? and keyword = ?",
-                         (ctx.guild.id, ctx.author.id, keyword))
+                         (guild_id, ctx.author.id, keyword))
 
         if check is not None:
             return await ctx.send(f"You already have this notification {self.emojis.get('hyunjinwtf')}")
 
-        db.execute("REPLACE INTO notifications values(?, ?, ?)", (ctx.guild.id, ctx.author.id, keyword))
-        await ctx.author.send(f"New notification for keyword `{keyword}` set in `{ctx.guild.name}` ")
-        await ctx.send(f"Set a notification! Check your DMs {self.emojis.get('vivismirk')}")
+        db.execute("REPLACE INTO notifications values(?, ?, ?)", (guild_id, ctx.author.id, keyword))
+        await ctx.author.send(f"New notification for keyword `{keyword}` set " + ("globally" if dm else f"in `{ctx.guild.name}`"))
+        if not dm:
+            await ctx.send(f"Set a notification!" + ("" if dm else f" Check your DMs {self.emojis.get('vivismirk')}"))
 
     @notification.command()
     async def remove(self, ctx, *, keyword):
         """Remove notification"""
-        await ctx.message.delete()
+        dm = ctx.guild is None
+        if dm:
+            guild_id = 0
+        else:
+            await ctx.message.delete()
+            guild_id = ctx.guild.id
 
         check = db.query("SELECT * FROM notifications WHERE guild_id = ? and user_id = ? and keyword = ?",
-                         (ctx.guild.id, ctx.author.id, keyword))
+                         (guild_id, ctx.author.id, keyword))
+
         if check is None:
             return await ctx.send(f"You don't have that notification {self.emojis.get('hyunjinwtf')}")
 
         db.execute("DELETE FROM notifications where guild_id = ? and user_id = ? and keyword = ?",
-                   (ctx.guild.id, ctx.author.id, keyword))
-        await ctx.author.send(f"Notification for keyword `{keyword}` removed for `{ctx.guild.name}` ")
-        await ctx.send(f"removed a notification! Check your DMs {self.emojis.get('vivismirk')}")
+                   (guild_id, ctx.author.id, keyword))
+        await ctx.author.send(f"Notification for keyword `{keyword}` removed " + ("globally" if dm else f"for `{ctx.guild.name}`"))
+        if not dm:
+            await ctx.send(f"removed a notification!" + ("" if dm else f" Check your DMs {self.emojis.get('vivismirk')}"))
 
     @notification.command()
     async def list(self, ctx):
@@ -108,19 +118,19 @@ class Notifications(commands.Cog):
 
         content = discord.Embed(title=":love_letter: Your notifications", color=discord.Color.red())
         for guild_id in guilds:
-            server = self.bot.get_guild(guild_id)
-            if server is None:
-                continue
-            content.add_field(name=server.name, value='\n'.join(f"└`{x}`" for x in guilds.get(guild_id)))
+            if guild_id == 0:
+                guild_name = "Global"
+            else:
+                server = self.bot.get_guild(guild_id)
+                if server is None:
+                    continue
+                guild_name = server.name
+
+            content.add_field(name=guild_name, value='\n'.join(f"└`{x}`" for x in guilds.get(guild_id)))
 
         await ctx.author.send(embed=content)
         if ctx.guild is not None:
             await ctx.send(f"List sent to your DMs {self.emojis.get('vivismirk')}")
-
-    @commands.command(hidden=True)
-    @commands.is_owner()
-    async def notifytest(self, ctx):
-        await send_notification(ctx.author, ctx.message)
 
 
 def setup(bot):
@@ -133,12 +143,10 @@ async def send_notification(user, message, pattern=None):
         name=f"{message.author}",
         icon_url=message.author.avatar_url
     )
-    if pattern is not None:
-        highlighted_text = re.sub(pattern, lambda x: f'**{x.group(0)}**', message.content)
-    else:
-        highlighted_text = message.content.replace(f">notifytest ", "")
+    highlighted_text = re.sub(pattern, lambda x: f'**{x.group(0)}**', message.content)
+    quoted_text = highlighted_text.replace('\n', '\n> ')
 
-    content.description = f"> {highlighted_text}\n[Go to message]({message.jump_url})"
+    content.description = f"{quoted_text}\n[Go to message]({message.jump_url})"
     content.set_footer(
         text=f"{message.guild.name} | #{message.channel.name}",
         icon_url=message.guild.icon_url
